feat(cockpit): update generative-ui example for streaming auto-detection
- Remove duplicate cockpit/langgraph/generative-ui/ (already exists under chat)
- Remove 'generative-ui' from langgraph manifest topics
- Replace old ChatGenerativeUiComponent usage with views() + [views] input
- Add WeatherCard, StatCard, Container view components
- Update system prompt to instruct LLM to return raw JSON specs
- Update docs guide to describe streaming pipeline
