+++ cockpit/chat/generative-ui/angular/src/app/views/container.component.ts
