+++ cockpit/chat/generative-ui/angular/src/app/views/stat-card.component.ts
