+++ cockpit/chat/generative-ui/angular/src/app/views/weather-card.component.ts
