+++ cockpit/chat/generative-ui/angular/src/environments/environment.development.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  langGraphApiUrl: 'http://localhost:4508/api',
-  streamingAssistantId: 'c-generative-ui',
+  langGraphApiUrl: 'http://localhost:4310/api',
+  generativeUiAssistantId: 'generative_ui',
 };
