--- cockpit/langgraph/generative-ui/angular/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@cacheplane/cockpit-langgraph-generative-ui-angular",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@cacheplane/chat": "^0.0.1",
-    "@cacheplane/render": "^0.0.1",
-    "@cacheplane/angular": "^0.0.1",
-    "@json-render/core": "^0.16.0",
-    "@langchain/langgraph-sdk": "^0.0.36"
-  },
-  "license": "PolyForm-Noncommercial-1.0.0",
-  "sideEffects": false
-}
