--- cockpit/langgraph/generative-ui/angular/project.json
@@ -1,54 +1,0 @@
-{
-  "name": "cockpit-langgraph-generative-ui-angular",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "cockpit/langgraph/generative-ui/angular/src",
-  "projectType": "application",
-  "targets": {
-    "build": {
-      "executor": "@angular/build:application",
-      "outputs": ["{options.outputPath.base}"],
-      "options": {
-        "outputPath": {
-          "base": "dist/cockpit/langgraph/generative-ui/angular",
-          "browser": ""
-        },
-        "browser": "cockpit/langgraph/generative-ui/angular/src/main.ts",
-        "tsConfig": "cockpit/langgraph/generative-ui/angular/tsconfig.app.json",
-        "styles": ["cockpit/langgraph/generative-ui/angular/src/styles.css"]
-      },
-      "configurations": {
-        "production": {
-          "budgets": [
-            { "type": "initial", "maximumWarning": "500kb", "maximumError": "1mb" },
-            { "type": "anyComponentStyle", "maximumWarning": "4kb", "maximumError": "8kb" }
-          ],
-          "outputHashing": "none"
-        },
-        "development": {
-          "optimization": false,
-          "extractLicenses": false,
-          "sourceMap": true,
-          "fileReplacements": [
-            {
-              "replace": "cockpit/langgraph/generative-ui/angular/src/environments/environment.ts",
-              "with": "cockpit/langgraph/generative-ui/angular/src/environments/environment.development.ts"
-            }
-          ]
-        }
-      },
-      "defaultConfiguration": "production"
-    },
-    "serve": {
-      "continuous": true,
-      "executor": "@angular/build:dev-server",
-      "configurations": {
-        "production": { "buildTarget": "cockpit-langgraph-generative-ui-angular:build:production" },
-        "development": { "buildTarget": "cockpit-langgraph-generative-ui-angular:build:development" }
-      },
-      "defaultConfiguration": "development",
-      "options": {
-        "proxyConfig": "cockpit/langgraph/generative-ui/angular/proxy.conf.json"
-      }
-    }
-  }
-}
