--- cockpit/langgraph/generative-ui/angular/proxy.conf.json
@@ -1,9 +1,0 @@
-{
-  "/api": {
-    "target": "http://localhost:8123",
-    "secure": false,
-    "changeOrigin": true,
-    "pathRewrite": { "^/api": "" },
-    "ws": true
-  }
-}
