--- cockpit/langgraph/generative-ui/angular/src/app/app.config.ts
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: PolyForm-Noncommercial-1.0.0
-import { ApplicationConfig } from '@angular/core';
-import { provideAgent } from '@cacheplane/angular';
-import { provideChat } from '@cacheplane/chat';
-import { environment } from '../environments/environment';
-
-export const appConfig: ApplicationConfig = {
-  providers: [
-    provideAgent({ apiUrl: environment.langGraphApiUrl }),
-    provideChat({}),
-  ],
-};
