--- cockpit/langgraph/generative-ui/angular/src/environments/environment.development.ts
@@ -1,5 +1,0 @@
-export const environment = {
-  production: false,
-  langGraphApiUrl: 'http://localhost:4310/api',
-  generativeUiAssistantId: 'generative_ui',
-};
