--- cockpit/langgraph/generative-ui/angular/src/environments/environment.ts
@@ -1,5 +1,0 @@
-export const environment = {
-  production: true,
-  langGraphApiUrl: '/api',
-  generativeUiAssistantId: 'generative_ui',
-};
