--- cockpit/langgraph/generative-ui/angular/src/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <title>LangGraph Generative UI — Angular</title>
-  <base href="/" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-950 text-gray-100 h-screen">
-  <app-generative-ui></app-generative-ui>
-</body>
-</html>
