--- cockpit/langgraph/generative-ui/angular/tsconfig.app.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../../../dist/out-tsc",
-    "lib": ["es2022", "dom"],
-    "types": [],
-    "emitDeclarationOnly": false
-  },
-  "files": ["src/main.ts"],
-  "include": ["src/**/*.ts"]
-}
