--- cockpit/langgraph/generative-ui/angular/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "extends": "../../../../tsconfig.base.json",
-  "compilerOptions": {
-    "noPropertyAccessFromIndexSignature": false,
-    "experimentalDecorators": true,
-    "module": "preserve",
-    "emitDeclarationOnly": false,
-    "composite": false,
-    "lib": ["es2022", "dom"],
-    "skipLibCheck": true,
-    "strict": false
-  },
-  "angularCompilerOptions": {
-    "enableI18nLegacyMessageIdFormat": false,
-    "strictInjectionParameters": false,
-    "strictInputAccessModifiers": false,
-    "strictTemplates": false
-  },
-  "files": [],
-  "include": [],
-  "references": [
-    { "path": "./tsconfig.app.json" }
-  ]
-}
