--- cockpit/langgraph/generative-ui/angular/vercel.json
@@ -1,6 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "buildCommand": "npx nx build cockpit-langgraph-generative-ui-angular",
-  "outputDirectory": "dist/cockpit/langgraph/generative-ui/angular/browser",
-  "framework": null
-}
