--- cockpit/langgraph/generative-ui/python/langgraph.json
@@ -1,8 +1,0 @@
-{
-  "graphs": {
-    "generative_ui": "./src/graph.py:graph"
-  },
-  "dependencies": ["."],
-  "python_version": "3.12",
-  "env": ".env"
-}
