--- cockpit/langgraph/generative-ui/python/project.json
@@ -1,24 +1,0 @@
-{
-  "name": "cockpit-langgraph-generative-ui-python",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "cockpit/langgraph/generative-ui/python/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{workspaceRoot}/dist/cockpit/langgraph/generative-ui/python"],
-      "options": {
-        "outputPath": "dist/cockpit/langgraph/generative-ui/python",
-        "main": "cockpit/langgraph/generative-ui/python/src/index.ts",
-        "tsConfig": "cockpit/langgraph/generative-ui/python/tsconfig.json"
-      }
-    },
-    "smoke": {
-      "executor": "nx:run-commands",
-      "options": {
-        "cwd": "cockpit/langgraph/generative-ui/python",
-        "command": "npx tsx -e \"import { langgraphGenerativeUiPythonModule } from './src/index.ts'; const m = langgraphGenerativeUiPythonModule; if (m.id !== 'langgraph-generative-ui-python') { throw new Error('Unexpected module: ' + m.id); }\""
-      }
-    }
-  }
-}
