--- cockpit/langgraph/generative-ui/python/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "../../../../dist/out-tsc"
-  },
-  "include": ["src/**/*.ts"]
-}
